requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastapi-security==0.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
passlib==1.7.4
prisma==0.15.0
pyasn1==0.4.8
pycparser==2.22
pydantic==1.10.21
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tomlkit==0.13.2
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
